raspy/requirements.txt
openai==1.3.0
picamera2==0.3.33
pyserial==3.5
python-dotenv==1.0.0
python-telegram-bot==22.5
replicate==0.20.0
RPi.GPIO==0.7.1a4